[core] move bit.h from compat, rename original bit_read -> bit_copy, add new bit_read which advances an input iterator
