+++ src/xci/data/BinaryWriter.h
@@ -9,7 +9,6 @@
 
 #include "BinaryBase.h"
 #include <xci/compat/endian.h>
-#include <xci/compat/bit.h>
 #include <xci/data/coding/leb128.h>
 #include <memory>
 #include <ostream>
